requirements.txt
django 
gunicorn 
djangorestframework 
corsheaders